Fix to virtual clock and application acceptance logic

+++ code/server/routes/utils/utils.js
@@ -11,11 +11,20 @@ const rearrangeProposals = async (req,res) => {
         let counterMovedProposals = 0;
         let expiredProposals = await dao.getExpiredProposals(selectedTimestamp);
         let proposalsToRevive = await dao.getProposalsToRevive(selectedTimestamp);
+        const acceptedProposalsIds = await dao.getAcceptedProposalsIds()
+            .then(ids => {
+                if(ids.error){
+                    return [];
+                } else {
+                    return ids;
+                }
+            });
+        console.log(acceptedProposalsIds)
         
         if(expiredProposals.length > 0 && !expiredProposals.error) {
             for (p of expiredProposals) { 
                 //archive expired proposals (insert into archived_thesis_proposals, delete from thesis proposals)
-                //also mark as cancelled all applications for expired proposals
+                //also mark as expired all applications for expired proposals
                 await dao.archiveProposal(new models.ThesisProposal(
                     p.id, //can be whatever, DB handles autoincrement id
                     p.title,
@@ -33,7 +42,7 @@ const rearrangeProposals = async (req,res) => {
                 ))
                     .then(() => {
                         dao.deleteProposal(p.id);
-                        dao.cancellPendingApplicationsForAThesis(p.id, p.supervisor.slice(0,6));
+                        dao.updateApplicationsForExpiredProposals(p.id, p.supervisor.slice(0,6));
                         counterMovedProposals++;
                     })
                     .catch( (e) => {
@@ -48,29 +57,33 @@ const rearrangeProposals = async (req,res) => {
                 //revive expired proposals (insert into thesis_proposals, delete from archived_thesis_proposals)
                 //put related applications statuses back to pending, but should also update proposal id since it's been
                 //inserted back as a new row in the table
-                await dao.saveNewProposal(new models.ThesisProposal(
-                    p.id, //can be whatever, DB handles autoincrement id
-                    p.title,
-                    p.supervisor,
-                    p.cosupervisors.join('-'),
-                    p.keywords,
-                    p.type,
-                    p.groups.join('-'),
-                    p.description,
-                    p.requirements,
-                    p.notes,
-                    p.expiration,
-                    p.level,
-                    p.cds.join(',')
-                ))
-                    .then((newId) => {
-                        dao.reviveExpiredApplications(p.id, newId)
-                            .then(dao.deleteProposalFromArchived(p.id));
-                        counterMovedProposals++;
-                    })
-                    .catch ( (e) => {
-                        return res.status(500).json(e.message);
-                    });
+                //NOTE: revive a proposal only if there isn't already an accepted application for that proposal, otherwise
+                //ignore it
+                if(!acceptedProposalsIds.includes(p.id)) {
+                    await dao.saveNewProposal(new models.ThesisProposal(
+                        p.id, //can be whatever, DB handles autoincrement id
+                        p.title,
+                        p.supervisor,
+                        p.cosupervisors.join('-'),
+                        p.keywords,
+                        p.type,
+                        p.groups.join('-'),
+                        p.description,
+                        p.requirements,
+                        p.notes,
+                        p.expiration,
+                        p.level,
+                        p.cds.join(',')
+                    ))
+                        .then((newId) => {
+                            dao.reviveExpiredApplications(p.id, newId)
+                                .then(dao.deleteProposalFromArchived(p.id));
+                            counterMovedProposals++;
+                        })
+                        .catch ( (e) => {
+                            return res.status(500).json(e.message);
+                        });
+                }
             }
         }
